Security - Renamed Authentication/Register Request/Resposne
