+++ src/main/java/com/alkemy/ong/auth/service/impl/UserAuthServiceImpl.java
@@ -3,7 +3,7 @@ package com.alkemy.ong.auth.service.impl;
 import com.alkemy.ong.auth.service.JwtUtil;
 import com.alkemy.ong.auth.service.UserAuthService;
 import com.alkemy.ong.model.request.AuthenticationRequest;
-import com.alkemy.ong.model.response.UserResponseLoginDTO;
+import com.alkemy.ong.model.response.AuthenticationResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.BadCredentialsException;
@@ -23,7 +23,7 @@ public class UserAuthServiceImpl implements UserAuthService {
     AuthenticationManager authManager;
 
     @Override
-    public UserResponseLoginDTO loginAttempt(AuthenticationRequest authenticationRequest) throws Exception {
+    public AuthenticationResponse loginAttempt(AuthenticationRequest authenticationRequest) throws Exception {
         UserDetails userDetails;
         try {
             UsernamePasswordAuthenticationToken newTry = new UsernamePasswordAuthenticationToken(
@@ -40,7 +40,7 @@ public class UserAuthServiceImpl implements UserAuthService {
 
         // Build Response:
         String jwt = jwtTokenUtil.generateToken(userDetails);
-        UserResponseLoginDTO foundUser =  UserResponseLoginDTO.builder()
+        AuthenticationResponse foundUser =  AuthenticationResponse.builder()
                 .email(userDetails.getUsername())
                 .password(userDetails.getPassword())
                 .jwt(jwt)
